--- src/backend/proxy.ts_
@@ -1,53 +1,0 @@
-import { ClientRequest } from 'http';
-
-import { NextFunction, Request, Response } from 'express';
-import { createProxyMiddleware } from 'http-proxy-middleware';
-import { v4 as uuidv4 } from 'uuid';
-
-import { Client, getOnBehalfOfAccessToken, stdoutLogger } from '@navikt/familie-backend';
-
-import { oboConfig, proxyUrl } from './config';
-
-const restream = (proxyReq: ClientRequest, req: Request, _res: Response) => {
-    if (req.body) {
-        const bodyData = JSON.stringify(req.body);
-        proxyReq.setHeader('Content-Type', 'application/json');
-        proxyReq.setHeader('Content-Length', Buffer.byteLength(bodyData));
-        proxyReq.write(bodyData);
-    }
-};
-
-const doMock = (path: string) => {
-    let mockPath = path.replace('oppgave/', 'mock/oppgave/');
-    mockPath = mockPath.replace('journalpost/', 'mock/journalpost/');
-    mockPath = mockPath.replace('feature', 'mock/feature');
-    mockPath = mockPath.replace('person', 'mock/person');
-    mockPath = mockPath.replace('fagsakForPerson', 'mock/fagsakForPerson');
-    return mockPath;
-};
-
-// eslint-disable-next-line
-export const doProxy: any = () => {
-    return createProxyMiddleware('/familie-ba-sak/api', {
-        changeOrigin: true,
-        logLevel: 'info',
-        onProxyReq: restream,
-        pathRewrite: (path: string, _req: Request) => {
-            const newPath = doMock(path.replace('/familie-ba-sak/api', ''));
-            return `/api${newPath}`;
-        },
-        secure: true,
-        target: `${proxyUrl}`,
-        logProvider: () => stdoutLogger,
-    });
-};
-
-export const attachToken = (authClient: Client) => {
-    return async (req: Request, _res: Response, next: NextFunction) => {
-        getOnBehalfOfAccessToken(authClient, req, oboConfig).then((accessToken: string) => {
-            req.headers['Nav-Call-Id'] = uuidv4();
-            req.headers.Authorization = `Bearer ${accessToken}`;
-            return next();
-        });
-    };
-};
