+++ src/frontend/index.less
@@ -21,9 +21,6 @@
 @import "./komponenter/Fagsak/Vilkårsvurdering/GeneriskVilkår/FastsettPeriode/fastsett-periode.less";
 @import "./komponenter/Fagsak/Totrinnskontroll/totrinnskontroll.less";
 
-@import "./komponenter/ManuellJournalføring/manuell-journalføring.less";
-
-
 // Felles
 @import "./komponenter/Felleskomponenter/Informasjonsbolk/informasjonsbolk.less";
 @import "./komponenter/Felleskomponenter/InputMedLabelTilVenstre/inputmedlabeltilvenstre.less";
