+++ src/frontend/komponenter/ManuellJournalføring/FeilDekoratør.tsx
