+++ src/frontend/komponenter/ManuellJournalføring/PdfDokument.tsx
