--- src/frontend/komponenter/ManuellJournalføring/manuell-journalføring.less
@@ -1,52 +1,0 @@
-.journalføring {
-
-    &.skjemasteg {
-        max-width: 50rem !important;
-    }
-
-    & .panel--gra {
-        max-width: 30rem;
-    }
-
-    height: calc(100vh - 3rem);
-    overflow: auto;
-
-    &__logisk-vedlegg {
-        display: flex;
-
-        &--input {
-            width: 20rem;
-        }
-
-        & .lukknapp {
-            margin-left: 1rem;
-            margin-top: 2rem;
-        }
-    }
-
-    &__opprett-behandling {
-        & .knapp.ikon-knapp {
-            margin-top: 1rem;
-            svg {
-                margin: 0.5rem;
-            }
-        }
-        & .typo-feilmelding {
-            margin-top: 0.5rem;
-        }
-    }
-
-
-    .behandlingliste {
-        &__tabell {
-            &--behandlingtype {
-                label {
-                    text-indent: 2rem !important;
-                }
-            }
-        }
-    }
-
-
-
-}
